requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
camelot-py[cv]==0.8.2     # via -r requirements.in
certifi==2020.12.5        # via requests
cffi==1.14.4              # via cryptography
chardet==4.0.0            # via camelot-py, pdfminer.six, requests
click==7.1.2              # via camelot-py, uvicorn
colorama==0.4.4           # via uvicorn
cryptography==3.3.1       # via pdfminer.six
et-xmlfile==1.0.1         # via openpyxl
fastapi==0.63.0           # via -r requirements.in
h11==0.12.0               # via uvicorn
idna==2.10                # via requests
jdcal==1.4.1              # via openpyxl
numpy==1.19.5             # via camelot-py, opencv-python, pandas
opencv-python==4.5.1.48   # via camelot-py
openpyxl==3.0.5           # via camelot-py
pandas==1.2.0             # via -r requirements.in, camelot-py
pdfminer.six==20201018    # via camelot-py
pycparser==2.20           # via cffi
pydantic==1.7.3           # via fastapi
pypdf2==1.26.0            # via camelot-py
python-dateutil==2.8.1    # via pandas
python-dotenv==0.15.0     # via -r requirements.in, uvicorn
pytz==2020.5              # via pandas
pyyaml==5.3.1             # via uvicorn
requests==2.25.1          # via -r requirements.in
simplejson==3.17.2        # via -r requirements.in
six==1.15.0               # via cryptography, python-dateutil
sortedcontainers==2.3.0   # via pdfminer.six
starlette==0.13.6         # via fastapi
urllib3==1.26.2           # via requests
uvicorn[standard]==0.13.3  # via -r requirements.in
watchgod==0.6             # via uvicorn
websockets==8.1           # via uvicorn
xlsxwriter==1.3.7         # via -r requirements.in
